innowise_task/requirements.txt
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
coverage==5.4
Django==3.1.6
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
factory-boy==3.2.0
Faker==6.1.1
geographiclib==1.50
geolocation-python==0.2.2
geopy==2.1.0
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
packaging==20.9
Pillow==8.1.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
PyJWT==2.0.1
pyparsing==2.4.7
pyserial==3.5
pytest==6.2.2
pytest-cov==2.11.1
pytest-django==4.1.0
pytest-factoryboy==2.1.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.3
